xml/parsing/manakai/entrefs-2.dat

#data
<p>&</p>
#errors
1;1;no XML decl;;;s
1;4;bare ero;;;m
#document
| <{} p>
|   "&"

#data
<p q="&<"/>
#errors
1;1;no XML decl;;;s
1;7;bare ero;;;m
1;8;lt in attr value;;;m
#document
| <{} p>
|   q="&<"

#data
<p>&&</p>
#errors
1;1;no XML decl;;;s
1;4;bare ero;;;m
1;5;bare ero;;;m
#document
| <{} p>
|   "&&"

#data
<p q="&&"/>
#errors
1;1;no XML decl;;;s
1;7;bare ero;;;m
1;8;bare ero;;;m
#document
| <{} p>
|   q="&&"

#data
<p q="&="/>
#errors
1;1;no XML decl;;;s
1;7;bare ero;;;m
#document
| <{} p>
|   q="&="

#data
<p q='&='/>
#errors
1;1;no XML decl;;;s
1;7;bare ero;;;m
#document
| <{} p>
|   q="&="

#data
<p q=&= />
#errors
1;1;no XML decl;;;s
1;6;unquoted attr value;;;m
1;6;bare ero;;;m
1;7;bad attribute value;;;m
#document
| <{} p>
|   q="&="

#data
<p q="&copy="/>
#errors
1;1;no XML decl;;;s
1;12;charref=;;;m
#document
| <{} p>
|   q="&copy="

#data
<p q="&copy&"/>
#errors
1;1;no XML decl;;;s
1;12;no refc;;;m
1;12;bare ero;;;m
1;7;entity not declared;;&copy;m
#document escaped
| <{} p>
|   q="\u00A9&"

#data
<p q='&copy='/>
#errors
1;1;no XML decl;;;s
1;12;charref=;;;m
#document
| <{} p>
|   q="&copy="

#data
<p q=&copy= />
#errors
1;1;no XML decl;;;s
1;6;unquoted attr value;;;m
1;11;bad attribute value;;;m
1;11;charref=;;;m
#document
| <{} p>
|   q="&copy="

#data
<p>& </p>
#errors
1;1;no XML decl;;;s
1;4;bare ero;;;m
#document
| <{} p>
|   "& "

#data escaped
<p>&\u0009</p>
#errors
1;1;no XML decl;;;s
1;4;bare ero;;;m
#document escaped
| <{} p>
|   "&\u0009"

#data
<p>&
#errors
1;1;no XML decl;;;s
1;4;bare ero;;;m
1;5;in body:#eof;;;m
#document
| <{} p>
|   "&"

#data
<p q="&
#errors
1;1;no XML decl;;;s
1;7;bare ero;;;m
1;8;unclosed attribute value;;;m
1;8;no root element;;;m
#document

#data
<p q="&"/>
#errors
1;1;no XML decl;;;s
1;7;bare ero;;;m
#document
| <{} p>
|   q="&"

#data
<p q='&'/>
#errors
1;1;no XML decl;;;s
1;7;bare ero;;;m
#document
| <{} p>
|   q="&"

#data
<p q="&'"/>
#errors
1;1;no XML decl;;;s
1;7;bare ero;;;m
#document
| <{} p>
|   q="&'"

#data
<p q='&"'/>
#errors
1;1;no XML decl;;;s
1;7;bare ero;;;m
#document
| <{} p>
|   q="&""

#data
<p q="& "/>
#errors
1;1;no XML decl;;;s
1;7;bare ero;;;m
#document
| <{} p>
|   q="& "

#data
<!DOCTYPE a[
<!ENTITY b "&&">
]><a/>
#errors
1;1;no XML decl;;;s
2;13;bare ero;;;m
2;14;bare ero;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY b "&&" "" "" >

#data
<!DOCTYPE a[
<!ENTITY b '&&'>
]><a/>
#errors
1;1;no XML decl;;;s
2;13;bare ero;;;m
2;14;bare ero;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY b "&&" "" "" >

#data
<!DOCTYPE a[
<!ENTITY b "&<">
]><a/>
#errors
1;1;no XML decl;;;s
2;13;bare ero;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY b "&<" "" "" >

#data
<!DOCTYPE a[
<!ENTITY b "&
#errors
1;1;no XML decl;;;s
2;13;bare ero;;;m
2;14;parser:EOF;;;m
2;14;parser:EOF;;;m
2;14;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a[
<!ENTITY b "& ">
]><a/>
#errors
1;1;no XML decl;;;s
2;13;bare ero;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities
| <!ENTITY b "& " "" "" >

#data escaped
<!DOCTYPE a[
<!ENTITY b "&\u0009">
]><a/>
#errors
1;1;no XML decl;;;s
2;13;bare ero;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities escaped
| <!ENTITY b "&\u0009" "" "" >

#data escaped
<!DOCTYPE a[
<!ENTITY b "&\u000a">
]><a/>
#errors
1;1;no XML decl;;;s
2;13;bare ero;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities escaped
| <!ENTITY b "&\u000a" "" "" >

#data escaped
<!DOCTYPE a[
<!ENTITY b "&\u000d">
]><a/>
#errors
1;1;no XML decl;;;s
2;13;bare ero;;;m
#document
| <!DOCTYPE a>
| <{} a>
#entities escaped
| <!ENTITY b "&\u000a" "" "" >

#data escaped
<!DOCTYPE a[
  <!ENTITY \u4e00\u4e01 "xxx">
]>
<a>&\u4e00\u4e01;</a>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
|   "xxx"

#data escaped
<!DOCTYPE a[
  <!ENTITY \u4e00\u4e01 "xxx">
]>
<a>&\u4e00\u4e01yy;</a>
#errors escaped
1;1;no XML decl;;;s
4;4;entity not declared;;\u4e00\u4e01yy;;m
#document escaped
| <!DOCTYPE a>
| <{} a>
|   "&\u4e00\u4e01yy;"

#data
<!DOCTYPE a[
  <!ENTITY ::: "yyy">
]>
<a>&:::;</a>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
|   "yyy"

#data
<p>&a&amp;</p>
#errors
1;1;no XML decl;;;s
1;4;bare ero;;;m
#document
| <{} p>
|   "&a&"

#data
<p>&a </p>
#errors
1;1;no XML decl;;;s
1;4;bare ero;;;m
#document
| <{} p>
|   "&a "

#data
<!DOCTYPE a[
<!ENTITY a<b "ccc">
]>
<p>&a<b;</p>
#errors
1;1;no XML decl;;;s
4;4;bare ero;;;m
4;11;nestc;;;m
4;12;no attr value;;;m
4;13;in body:#eof;;;m
#document
| <!DOCTYPE a>
| <{} p>
|   "&a"
|   <{} b;<>
|     p=""
#entities
<!ENTITY a<b "ccc" "" "" >

#data escaped
<!DOCTYPE a[
  <!ENTITY b "&\u4e00\u4e01;">
]>
<a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>
#entities escaped
<!ENTITY b "&\u4e00\u4e01;" "" "" >
